Updates:
1. Handle use-case when a user already is registered, send out a greeting
2. Fix the issue where new users were not able to register.

+++ src/main/java/com/kajjoy/repo/UserRepository.java
@@ -4,7 +4,9 @@ import com.kajjoy.model.AuthenticatedUser;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface UserRepository extends MongoRepository<AuthenticatedUser,Long> {
+public interface UserRepository extends MongoRepository<AuthenticatedUser,BigInteger> {
     AuthenticatedUser findByUsername(String username);
 }
